Module vanetza_sys::vanetza

Modules

aid

btp

dcc

geonet

security

Structs

ByteBufferConvertible

ByteBufferConvertible is an extensible mechanism for providing ByteBuffer representations of various data structures and objects.

ChunkPacket

\brief ChunckPacket is a packet consisting of several memory chunks

CohesivePacket

A cohesive packet is stored in contiguous memory

InputArchive

This is a drop-in replacement for boost::archive::binary_iarchive

MacAddress

OutputArchive

This is a drop-in replacement for boost::archive::binary_oarchive

PacketVariant

PositionFix

Runtime

Runtime provides current time and enables scheduling of tasks for later execution.

Enums

OsiLayer

Type Aliases

ByteBuffer

DownPacket

\brief ChunckPacket is a packet consisting of several memory chunks

ItsAid