Module vanetza_sys::vanetza::geonet

Structs

Address

Area

autocxx bindings couldn’t be generated: This item name is used in multiple namespaces. At present, autocxx and cxx allow only one type of a given name. This limitation will be fixed in future. (Items found with this name: vanetza::units::Area, vanetza::geonet::Area)

CartesianPosition

Cartesian position. As it is not strictly specified in any GeoNet standard document, I define it as point in an ENU coordinate system and its units as meters.

CbfPacketBuffer

CbfPacketBuffer facilitates implementation of contention based forwarding

Circle

DataConfirm

DataIndication

DataRequest

autocxx bindings couldn’t be generated: This item name is used in multiple namespaces. At present, autocxx and cxx allow only one type of a given name. This limitation will be fixed in future. (Items found with this name: vanetza::geonet::DataRequest, vanetza::dcc::DataRequest)

DataRequestWithAddress

DataRequestWithArea

DccFieldGenerator

Ellipse

GacDataRequest

GbcDataRequest

GeodeticPosition

GucDataRequest

Lifetime

LocationTable

GeoNetworking LocationTable See section 7.1 of EN 302 636-4-1 for details.

LongPositionVector

ManagementInformationBase

The Management Information Base (MIB) defines the GeoNetworking protocol constants.

Rectangle

Router

Router is the central entity for GeoNet communication

ShbDataRequest

ShortPositionVector

TrafficClass

TransportInterface

TsbDataRequest

inside_or_at_border

autocxx bindings couldn’t be generated: This item depends on some other type(s) which autocxx could not generate, some of them are: vanetza::geonet::Area

Enums

AddrConfMethod

BroadcastForwarding

DataConfirm_ResultCode

InterfaceType

SecurityDecapHandling

StationType

TransportType

UnicastForwarding

UpperProtocol

Functions

geometric_function

geometric_function1

geometric_function2

local_cartesian

Derive cartesian position ENU from geodetic WGS84 coordinates and a WGS84 reference point which becomes the cartesian origin. \param origin WGS84 reference point becoming origin \param position Calculate cartesian coordinates for this point \return Cartesian coordinates of position relative to origin

Type Aliases

MIB

The Management Information Base (MIB) defines the GeoNetworking protocol constants.

Router_DownPacketPtr