Struct vanetza_sys::vanetza::geonet::Router

#[repr(C, align(8))]
pub struct Router { /* private fields */ }

Router is the central entity for GeoNet communication

Incoming and outgoing GeoNet packets are handled by the router. It may even dispatch own packets (beacons) if necessary.

This implementation follows EN 302 636-4-1 v1.3.1

Implementations

impl Router

pub fn request( self: Pin<&mut Router>, arg1: &ShbDataRequest, arg2: UniquePtr<ChunkPacket> ) -> DataConfirm

\brief Request to send payload per single hop broadcast (SHB). If security is enabled, the message gets encapsulated in a security envelope. Returns whether data was valid to be sent.

\param request \param payload from upper layers \return result code if packet has been accepted

pub fn request1( self: Pin<&mut Router>, arg1: &GbcDataRequest, arg2: UniquePtr<ChunkPacket> ) -> DataConfirm

\brief Request to send payload per GeoBroadcast (GBC). If security is enabled, the message gets encapsulated in a security envelope. Returns whether data was valid to be sent.

\param request \param payload from upper layers \return result code if packet has been accepted

pub fn request2( self: Pin<&mut Router>, arg1: &GucDataRequest, arg2: UniquePtr<ChunkPacket> ) -> DataConfirm

pub fn request3( self: Pin<&mut Router>, arg1: &GacDataRequest, arg2: UniquePtr<ChunkPacket> ) -> DataConfirm

pub fn request4( self: Pin<&mut Router>, arg1: &TsbDataRequest, arg2: UniquePtr<ChunkPacket> ) -> DataConfirm

pub fn new<'a>( arg1: Pin<&'a mut Runtime>, arg2: &'a ManagementInformationBase ) -> impl New<Output = Self> + 'a

impl Router

pub fn update_position(self: Pin<&mut Self>, arg1: &PositionFix)

\brief Update router’s local position vector

\param fix current position fix

impl Router

pub unsafe fn set_transport_handler( self: Pin<&mut Self>, proto: UpperProtocol, ifc: *mut TransportInterface )

\brief Register a transport protocol handler.

\param proto register handler for this upper protocol \param ifc use this interface or disable handling if nullptr

impl Router

pub unsafe fn set_security_entity( self: Pin<&mut Self>, entity: *mut SecurityEntity )

\brief Register security entity used when itsGnSecurity is enabled

\param entity security entity

impl Router

pub unsafe fn set_access_interface( self: Pin<&mut Self>, ifc: *mut RequestInterface )

\brief Register access layer interface

\param ifc interface used for passing packets down to access layer

impl Router

pub unsafe fn set_dcc_field_generator( self: Pin<&mut Self>, dcc: *mut DccFieldGenerator )

\brief Register generator for DCC-MCO fields

\param dcc DCC-MCO field generator or nullptr for disabling feature

impl Router

pub fn set_address(self: Pin<&mut Self>, arg1: &Address)

\brief Set Router’s own GeoNetworking address

\param addr

impl Router

pub fn get_mib(&self) -> &ManagementInformationBase

\brief Get Management Information Base (MIB) \return read-only reference to MIB

impl Router

pub fn get_cbf_buffer(&self) -> &CbfPacketBuffer

\brief Get the Contention-Based-Forwarding buffer

\return read-only reference to CBF packet buffer

impl Router

pub fn get_location_table(&self) -> &LocationTable

\brief Get the LocationTable. The table holds information about neighbouring ITS-Routers.

\return read-only reference to LocationTable

impl Router

pub fn get_local_position_vector(&self) -> &LongPositionVector

\brief Get the local position vector. This vector describes the current position of the router.

\return read-only reference to LongPositionVector

impl Router

pub fn outside_sectorial_contention_area( &self, sender: &MacAddress, forwarder: &MacAddress ) -> bool

\brief Check if router is outside the sectorial contention area See TS 102 636-4-1 v1.2.3 section E.4 and figure E.2 for details.

\param sender \param forwarder \return bool true if either sender or forwarder is outside

impl Router

pub fn set_random_seed(self: Pin<&mut Self>, seed: c_ulong)

\brief Set seed for internal random number generator (RNG) RNG is used e.g. for random Beacon jitter

\param seed reset RNG’s state to this seed